+++ test/tint/unittest/reader/spirv/ImageQueryLevels_SignedResult_SpvParserHandleTest_SampledImageAccessTest_Variable_4.spvasm.expected.glsl
@@ -22,7 +22,7 @@ void main_1() {
   vec2 coords12 = vf12;
   vec3 coords123 = vf123;
   vec4 coords1234 = vf1234;
-  int x_99 = textureQueryLevels(x_20_1);
+  int x_99 = int(uint(textureQueryLevels(x_20_1)));
   return;
 }
